requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
Flask-Migrate
PyMySQL
qrcode[pil]
python-dotenv